requirements.txt
asgiref==3.9.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Django==5.2.5
djangorestframework==3.16.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
